requirements.txt
requests>=2.31,<3.0
python-telegram-bot>=21.0,<22.0
python-dotenv>=1.0,<2.0
redis>=5.0,<6.0
